requirements.txt
spotipy==2.22.1
mido==1.2.10
python-dotenv==0.21.1